+++ lib/model/place_details.dart
@@ -5,18 +5,7 @@ class PlaceDetails {
   PlaceDetails({this.result, this.status});
 
   PlaceDetails.fromJson(Map<String, dynamic> json) {
-    result = json['result'] != null ? Result.fromJson(json['result']) : null;
-    status = json['status'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-
-    if (result != null) {
-      data['result'] = result!.toJson();
-    }
-    data['status'] = status;
-    return data;
+    result = Result.fromJson(json);
   }
 }
 
@@ -37,77 +26,48 @@ class Result {
   String? vicinity;
   String? website;
 
-  Result(
-      {this.addressComponents,
-      this.adrAddress,
-      this.formattedAddress,
-      this.geometry,
-      this.icon,
-      this.name,
-      this.photos,
-      this.placeId,
-      this.reference,
-      this.scope,
-      this.types,
-      this.url,
-      this.utcOffset,
-      this.vicinity,
-      this.website});
+  Result({
+    this.addressComponents,
+    this.adrAddress,
+    this.formattedAddress,
+    this.geometry,
+    this.icon,
+    this.name,
+    this.photos,
+    this.placeId,
+    this.reference,
+    this.scope,
+    this.types,
+    this.url,
+    this.utcOffset,
+    this.vicinity,
+    this.website,
+  });
 
   Result.fromJson(Map<String, dynamic> json) {
-    if (json['address_components'] != null) {
+    if (json['addressComponents'] != null) {
       addressComponents = [];
-      json['address_components'].forEach((v) {
+      json['addressComponents'].forEach((v) {
         addressComponents!.add(AddressComponents.fromJson(v));
       });
     }
-    adrAddress = json['adr_address'];
-    formattedAddress = json['formatted_address'];
-    geometry =
-        json['geometry'] != null ? Geometry.fromJson(json['geometry']) : null;
-    icon = json['icon'];
+    adrAddress = json['adrFormatAddress'];
+    formattedAddress = json['formattedAddress'];
+    icon = json['iconMaskBaseUri'];
     name = json['name'];
+    geometry =
+        json['location'] != null ? Geometry.fromJson(json['location']) : null;
     if (json['photos'] != null) {
       photos = [];
       json['photos'].forEach((v) {
         photos!.add(Photos.fromJson(v));
       });
     }
-    placeId = json['place_id'];
-    reference = json['reference'];
-    scope = json['scope'];
+    placeId = json['placeId'];
     types = json['types'].cast<String>();
-    url = json['url'];
-    utcOffset = json['utc_offset'];
-    vicinity = json['vicinity'];
-    website = json['website'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    if (addressComponents != null) {
-      data['address_components'] =
-          addressComponents!.map((v) => v.toJson()).toList();
-    }
-    data['adr_address'] = adrAddress;
-    data['formatted_address'] = formattedAddress;
-    if (geometry != null) {
-      data['geometry'] = geometry!.toJson();
-    }
-    data['icon'] = icon;
-    data['name'] = name;
-    if (photos != null) {
-      data['photos'] = photos!.map((v) => v.toJson()).toList();
-    }
-    data['place_id'] = placeId;
-    data['reference'] = reference;
-    data['scope'] = scope;
-    data['types'] = types;
-    data['url'] = url;
-    data['utc_offset'] = utcOffset;
-    data['vicinity'] = vicinity;
-    data['website'] = website;
-    return data;
+    url = json['googleMapsUri'];
+    utcOffset = json['utcOffsetMinutes'];
+    website = json['websiteUri'];
   }
 }
 
@@ -123,14 +83,6 @@ class AddressComponents {
     shortName = json['short_name'];
     types = json['types'].cast<String>();
   }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    data['long_name'] = longName;
-    data['short_name'] = shortName;
-    data['types'] = types;
-    return data;
-  }
 }
 
 class Geometry {
@@ -140,21 +92,7 @@ class Geometry {
   Geometry({this.location, this.viewport});
 
   Geometry.fromJson(Map<String, dynamic> json) {
-    location =
-        json['location'] != null ? Location.fromJson(json['location']) : null;
-    viewport =
-        json['viewport'] != null ? Viewport.fromJson(json['viewport']) : null;
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    if (location != null) {
-      data['location'] = location!.toJson();
-    }
-    if (viewport != null) {
-      data['viewport'] = viewport!.toJson();
-    }
-    return data;
+    location = Location.fromJson(json);
   }
 }
 
@@ -165,15 +103,8 @@ class Location {
   Location({this.lat, this.lng});
 
   Location.fromJson(Map<String, dynamic> json) {
-    lat = json['lat'];
-    lng = json['lng'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    data['lat'] = lat;
-    data['lng'] = lng;
-    return data;
+    lat = json['latitude'];
+    lng = json['longitude'];
   }
 }
 
@@ -189,17 +120,6 @@ class Viewport {
     southwest =
         json['southwest'] != null ? Location.fromJson(json['southwest']) : null;
   }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    if (northeast != null) {
-      data['northeast'] = northeast!.toJson();
-    }
-    if (southwest != null) {
-      data['southwest'] = southwest!.toJson();
-    }
-    return data;
-  }
 }
 
 class Photos {
@@ -211,18 +131,9 @@ class Photos {
   Photos({this.height, this.htmlAttributions, this.photoReference, this.width});
 
   Photos.fromJson(Map<String, dynamic> json) {
-    height = json['height'];
-    htmlAttributions = json['html_attributions'].cast<String>();
-    photoReference = json['photo_reference'];
-    width = json['width'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    data['height'] = height;
-    data['html_attributions'] = htmlAttributions;
-    data['photo_reference'] = photoReference;
-    data['width'] = width;
-    return data;
+    height = json['heightPx'];
+    htmlAttributions = json['authorAttributions'].cast<String>();
+    photoReference = json['name'];
+    width = json['widthPx'];
   }
 }